simple_image_machine.c,h ver 0.5_alt
Located source of bug causing random seg faults: failed to malloc memory for the creation of a template
structure from inside the function loadTemplate.  Returned a pointer to a template structure that lost
its memory.  Once the structure had memory allocated for it, all seg faults disappeared.  Fixed up
little bits of code that were incorrect in dealing with the buffers, but basically the code worked as
was after that bug was fixed